fix #3 Changed Transaction Amount type from Long to BigDecimal

+++ transaction/src/main/java/ee/qrental/transaction/application/port/in/query/GetBalanceQuery.java
@@ -2,11 +2,12 @@ package ee.qrental.transaction.application.port.in.query;
 
 import ee.qrental.transaction.application.port.in.response.balance.BalanceResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface GetBalanceQuery {
     List<BalanceResponse> getAll();
 
-    Long getTotalByDriverId(final Long driverId);
+    BigDecimal getTotalByDriverId(final Long driverId);
 
 }
