+++ transaction/src/main/java/ee/qrental/transaction/application/port/in/response/transaction/TransactionResponse.java
@@ -3,6 +3,7 @@ package ee.qrental.transaction.application.port.in.response.transaction;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -12,7 +13,7 @@ public class TransactionResponse {
     private String type;
     private String driverInfo;
     private Integer callSign;
-    private Long realAmount;
+    private BigDecimal realAmount;
     private Integer weekNumber;
     private LocalDate date;
     private LocalDate dateStamp;
